Se terminaron las impl, faltan sus respectivos controller

+++ src/main/java/org/steamclone/dtos/TransactionDetailDTO.java
@@ -1,12 +1,8 @@
 package org.steamclone.dtos;
 
-import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
-import lombok.EqualsAndHashCode;
 import lombok.Getter;
 import lombok.Setter;
-import org.steamclone.models.entities.Game;
-import org.steamclone.models.entities.Transaction;
 
 @AllArgsConstructor
 @Getter
@@ -14,7 +10,7 @@ import org.steamclone.models.entities.Transaction;
 public class TransactionDetailDTO {
 
     private int id;
-    private float price;
+    private double price;
     private int transaction;
     private int game;
